requirements.txt
asgiref==3.3.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
certifi==2020.4.5.1
cffi==1.14.0
channels==3.0.3
chardet==3.0.4
cmake==3.18.4.post1
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.5
django-bootstrap-breadcrumbs==0.9.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==12.0.3
django-cas-ng==	4.1.1
django-ckeditor==5.9.0
django-common-helpers==0.9.2
django-cors-headers==3.4.0
django-cron==0.5.1
django-debug-toolbar==2.2
django-discover-runner==1.0
django-import-export==2.0.2
django-js-asset==1.2.2
django-multiselectfield==0.1.12
django-redis==4.12.1
django-render-block==0.7
django-role-permissions==2.2.1
django-slack==5.14.3
django-storages==1.9.1
django-templated-email==2.3.0
django-widget-tweaks==1.4.8
easy-thumbnails==2.7.1
et-xmlfile==1.0.1
hyperlink==21.0.0
idna==2.9
img2pdf==0.4.0
incremental==17.5.0
jdcal==1.4.1
lxml==4.5.0
MarkupPy==1.14
mysqlclient==1.4.6
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
pdf2image==1.14.0
pikepdf==2.1.1
Pillow==7.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.1
pytesseract==0.3.6
python-dotenv==0.17.1
python-docx==0.8.10
python3-memcached==1.51
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
redis==3.5.3
reportlab==3.5.42
requests==2.23.0
requests-oauthlib==1.3.0
sentry-sdk==0.20.3
service-identity==18.1.0
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
tablib==1.1.0
Twisted==20.3.0
txaio==20.12.1
urllib3==1.25.9
Wand==0.6.3
whitenoise==5.1.0
xlrd==1.2.0
xlwt==1.3.0
zope.interface==5.2.0
